Desenvolvimento agenda de dirigentes, área administrativa.

+++ administrator/components/com_agendadirigentes/views/cargos/view.html.php
@@ -70,14 +70,29 @@ class AgendaDirigentesViewCargos extends JViewLegacy
          */
         protected function addToolBar( $total = NULL ) 
         {
-                JToolBarHelper::title(JText::_('COM_AGENDADIRIGENTES_MANAGER_CARGOS').
-                        //Reflect number of items in title!
-                        ($total?' <span style="font-size: 0.5em; vertical-align: middle;">('.$total.')</span>':'')
-                        , 'compromisso');
+            JToolBarHelper::title(JText::_('COM_AGENDADIRIGENTES_MANAGER_CARGOS').
+                    //Reflect number of items in title!
+                    ($total?' <span style="font-size: 0.5em; vertical-align: middle;">('.$total.')</span>':'')
+                    , 'compromisso');
+
+            if ($this->canDo->get('core.create'))
                 JToolBarHelper::addNew('cargo.add');
+
+            if (($this->canDo->get('core.edit')) || ($this->canDo->get('core.edit.own')))
                 JToolBarHelper::editList('cargo.edit');
+
+            if ($this->canDo->get('core.edit.state'))
+            {
+                JToolBarHelper::publishList('cargos.publish');
+                JToolBarHelper::unpublishList('cargos.unpublish');
+            }
+
+            if ($this->canDo->get('core.delete'))
+            {
                 JToolBarHelper::deleteList('', 'cargos.delete');
-                JToolBarHelper::preferences('com_agendadirigentes');
+            }
+
+            JToolBarHelper::preferences('com_agendadirigentes');
         }
 
         /**
